+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/status/TenantStatusDto.java
@@ -44,16 +44,18 @@ public class TenantStatusDto {
         // @NotNull(message = "instance can not be null")
         public List<@Valid EventStatusDto> jobs;
 
-        @JsonProperty("jobs")
-        public void setJobs(List<EventStatusDto> jobs) {
-                if (jobs == null) {
-                        this.jobs = null;
-                        return;
-                }
+//        @JsonProperty("jobs")
+//        public void setJobs(List<EventStatusDto> jobs) {
+//                if (jobs == null) {
+//                        this.jobs = null;
+//                        return;
+//                }
+//
+//                // Filter out CHECK_READINESS
+//                this.jobs = jobs.stream()
+//                        .filter(job -> !"CHECK_READINESS".equalsIgnoreCase(job.getName()))
+//                        .toList();
+//        }
 
-                // Filter out CHECK_READINESS
-                this.jobs = jobs.stream()
-                        .filter(job -> !"CHECK_READINESS".equalsIgnoreCase(job.getName()))
-                        .toList();
-        }
+
 }